2025/06/10 🔳MacStudio Web端 ❶:碎课拼成整课并转换成正课，❷:零碎加课已经转换正课一览和撤销正课还原到零碎加课功能

+++ 01FlutterApps/middle_layer_springboot/knpiano_java/src/main/java/com/liu/springboot04web/bean/Kn02F004UnpaidBean.java
@@ -36,7 +36,7 @@ public class Kn02F004UnpaidBean implements KnPianoBean {
     protected String    subjectSubName;
     protected String    lsnMonth;
     protected Integer   payStyle;
-    protected Integer   lsnCount;
+    protected float     lsnCount;
     protected float     lsnFee;
     protected Integer   ownFlg;
     // 为了按月交费的计划课的精算业务，需要《学生档案》表里对象科目的最新价格
@@ -126,10 +126,10 @@ public class Kn02F004UnpaidBean implements KnPianoBean {
     public void setPayStyle(Integer payStyle) {
         this.payStyle = payStyle;
     }
-    public Integer getLsnCount() {
+    public float getLsnCount() {
         return lsnCount;
     }
-    public void setLsnCount(Integer lsnCount) {
+    public void setLsnCount(float lsnCount) {
         this.lsnCount = lsnCount;
     }
     public String getLsnMonth() {
